Add files via upload
This folder contains complete template which is ready for fulfilling React components and Node.js functions. I suggest that when you think you finish a part completely (front-end and back-end), you may add your related code into this template and try to run it to see if there are some problems. In the end, this folder will be a deliverable project code.